Consolidate SDImageTransform and SDXLImageTransform. Previously, SDXLImageTransform was equivalent in behaviour, but just returned some additional information.
