+++ src/invoke_training/config/shared/data/data_loader_config.py
@@ -5,7 +5,6 @@ from pydantic import BaseModel
 from invoke_training.config.shared.data.dataset_config import ImageCaptionDatasetConfig, ImageDirDatasetConfig
 from invoke_training.config.shared.data.transform_config import (
     SDImageTransformConfig,
-    SDXLImageTransformConfig,
     ShuffleCaptionTransformConfig,
     TextualInversionCaptionConfig,
 )
@@ -25,8 +24,6 @@ class ImageCaptionSDDataLoaderConfig(BaseModel):
 
 class ImageCaptionSDXLDataLoaderConfig(ImageCaptionSDDataLoaderConfig):
     type: Literal["IMAGE_CAPTION_SDXL_DATA_LOADER"] = "IMAGE_CAPTION_SDXL_DATA_LOADER"
-
-    image_transforms: SDXLImageTransformConfig
 
 
 class DreamboothSDDataLoaderConfig(BaseModel):
@@ -52,8 +49,6 @@ class DreamboothSDDataLoaderConfig(BaseModel):
 class DreamboothSDXLDataLoaderConfig(DreamboothSDDataLoaderConfig):
     type: Literal["DREAMBOOTH_SDXL_DATA_LOADER"] = "DREAMBOOTH_SDXL_DATA_LOADER"
 
-    image_transforms: SDXLImageTransformConfig
-
 
 class TextualInversionSDDataLoaderConfig(BaseModel):
     type: Literal["TEXTUAL_INVERSION_SD_DATA_LOADER"] = "TEXTUAL_INVERSION_SD_DATA_LOADER"
@@ -77,7 +72,3 @@ class TextualInversionSDDataLoaderConfig(BaseModel):
 
 class TextualInversionSDXLDataLoaderConfig(TextualInversionSDDataLoaderConfig):
     type: Literal["TEXTUAL_INVERSION_SDXL_DATA_LOADER"] = "TEXTUAL_INVERSION_SDXL_DATA_LOADER"
-
-    image_transforms: SDXLImageTransformConfig
-    """The image transforms to apply to all images.
-    """
