+++ src/invoke_training/training/shared/data/data_loaders/dreambooth_sdxl_dataloader.py
@@ -18,7 +18,7 @@ from invoke_training.training.shared.data.datasets.transform_dataset import Tran
 from invoke_training.training.shared.data.transforms.constant_field_transform import ConstantFieldTransform
 from invoke_training.training.shared.data.transforms.drop_field_transform import DropFieldTransform
 from invoke_training.training.shared.data.transforms.load_cache_transform import LoadCacheTransform
-from invoke_training.training.shared.data.transforms.sdxl_image_transform import SDXLImageTransform
+from invoke_training.training.shared.data.transforms.sd_image_transform import SDImageTransform
 from invoke_training.training.shared.data.transforms.tensor_disk_cache import TensorDiskCache
 
 
@@ -78,7 +78,7 @@ def build_dreambooth_sdxl_dataloader(
     all_transforms = []
     if vae_output_cache_dir is None:
         all_transforms.append(
-            SDXLImageTransform(
+            SDImageTransform(
                 resolution=data_loader_config.image_transforms.resolution,
                 center_crop=data_loader_config.image_transforms.center_crop,
                 random_flip=data_loader_config.image_transforms.random_flip,
