+++ tests/invoke_training/training/shared/data/data_loaders/test_dreambooth_sdxl_dataloader.py
@@ -2,7 +2,7 @@ import torch
 
 from invoke_training.config.shared.data.data_loader_config import DreamboothSDXLDataLoaderConfig
 from invoke_training.config.shared.data.dataset_config import ImageDirDatasetConfig
-from invoke_training.config.shared.data.transform_config import SDXLImageTransformConfig
+from invoke_training.config.shared.data.transform_config import SDImageTransformConfig
 from invoke_training.training.shared.data.data_loaders.dreambooth_sdxl_dataloader import (
     build_dreambooth_sdxl_dataloader,
 )
@@ -18,7 +18,7 @@ def test_build_dreambooth_sdxl_dataloader(image_dir):  # noqa: F811
         class_caption="test class prompt",
         # For testing, we just use the same directory for the instance and class datasets.
         class_dataset=ImageDirDatasetConfig(dataset_dir=str(image_dir)),
-        image_transforms=SDXLImageTransformConfig(resolution=512),
+        image_transforms=SDImageTransformConfig(resolution=512),
     )
     data_loader = build_dreambooth_sdxl_dataloader(data_loader_config=config, batch_size=2)
 
@@ -53,7 +53,7 @@ def test_build_dreambooth_sdxl_dataloader_no_class_dataset(image_dir):  # noqa: 
     config = DreamboothSDXLDataLoaderConfig(
         instance_caption="test instance prompt",
         instance_dataset=ImageDirDatasetConfig(dataset_dir=str(image_dir)),
-        image_transforms=SDXLImageTransformConfig(resolution=512),
+        image_transforms=SDImageTransformConfig(resolution=512),
     )
     data_loader = build_dreambooth_sdxl_dataloader(
         data_loader_config=config,
